backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
python-multipart>=0.0.6
email-validator>=2.0.0

